DASH: Avoid rounding error in getSegmentCount
Issue: #8804
PiperOrigin-RevId: 369484117

+++ library/dash/src/main/java/com/google/android/exoplayer2/source/dash/DashSegmentIndex.java
@@ -89,7 +89,7 @@ public interface DashSegmentIndex {
    *     C#TIME_UNSET} if the period's duration is not yet known.
    * @return The number of segments in the index, or {@link #INDEX_UNBOUNDED}.
    */
-  int getSegmentCount(long periodDurationUs);
+  long getSegmentCount(long periodDurationUs);
 
   /**
    * Returns the number of available segments in the index.
@@ -99,7 +99,7 @@ public interface DashSegmentIndex {
    * @param nowUnixTimeUs The current time in milliseconds since the Unix epoch.
    * @return The number of available segments in the index.
    */
-  int getAvailableSegmentCount(long periodDurationUs, long nowUnixTimeUs);
+  long getAvailableSegmentCount(long periodDurationUs, long nowUnixTimeUs);
 
   /**
    * Returns the time, in microseconds, at which a new segment becomes available, or {@link
